Fix: nreplace shouldn't proceed from right to left
The buffer based version of nreplace failed to comply with
the documentation that nreplace should replace from right to
left. Since I can't think of a way to fix this elegantly while
keeping the speed improvement brought by previous commit, I
changed the doc instead, which may break some code.

If retrocompatibility on such issue is deemed important, then
we may add an additional parameter to choose direction, or at
least provide a nreplace_rev so that code relying on former
behavior can easily be adapted?